requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.6.4
psycopg2-binary==2.9.9
nebula3-python==3.5.0
qdrant-client==1.7.0
celery[redis]==5.3.6
redis==5.0.3
openai==1.14.3
tenacity==8.2.3
python-dotenv==1.0.1
asyncpg==0.29.0
pgvector==0.2.4
pytest==8.0.0
websockets==12.0 